Move useMongoTimestamp tests to be part of timestamps
